refactor: add `actions/index.ts`
